github_integration_service/requirements.txt
flask
flask-cors
pdfkit
nltk
requests
python-dotenv
flask_sqlalchemy
gunicorn
GitPython
